requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
python-dotenv
pydantic
langchain-groq
scikit-learn
python-multipart